% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_area.R
\name{get_area}
\alias{get_area}
\title{Get Ameba Area}
\usage{
get_area(routes_df_xy, max_km, concavity = 2, len_th = 0)
}
\arguments{
\item{routes_df_xy}{A data frame containing route information with x and y coordinates.}

\item{max_km}{The maximum distance threshold in kilometers for including routes in the area.}

\item{concavity}{The concavity parameter for the concaveman algorithm.}

\item{len_th}{The length threshold parameter for the concaveman algorithm.}
}
\value{
A sf object representing the concave hull area.
}
\description{
Create the concave area
}
