% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_routes_inside.R
\name{get_routes_inside}
\alias{get_routes_inside}
\title{Get Routes Inside}
\usage{
get_routes_inside(
  lat,
  lon,
  radius_km = 5,
  grid_size = 1,
  max_km = NULL,
  crs_ = 4326,
  f_grid = ceiling,
  .progress = T
)
}
\arguments{
\item{lat}{The latitude of the initial point.}

\item{lon}{The longitude of the initial point.}

\item{radius_km}{The radius in kilometers to define the area within which the routes are calculated. Default is 5 kilometers.}

\item{grid_size}{The size of the grid cells in kilometers. Default is 1 kilometer.}

\item{max_km}{The maximum distance in kilometers allowed for the routes. If specified, routes beyond this distance will be filtered out. Default is NULL (no maximum distance).}

\item{crs_}{The coordinate reference system (CRS) code to use. Default is 4326 (WGS84).}

\item{f_grid}{The function to round up the number of grid cells. Default is ceiling.}

\item{.progress}{Logical value indicating whether to display a progress bar. Default is TRUE.}
}
\value{
A spatial data frame containing the routes inside the specified radius.
}
\description{
Calculates routes inside a
specified radius around a given latitude
and longitude point. It divides the area
into a grid of smaller polygons and computes the routes from the initial point to the centroids of each grid cell within the radius.
}
\examples{
get_routes_inside(lat = -12.0464, lon = -77.0428, radius_km = 5)
}
